feat: 로그인 성공

+++ app/src/main/java/com/hongul/fliq/model/card/Card.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.Immutable
 @Immutable
 data class Card(
     val id: Int,
-    val owner: Int,
+    val owner: Long,
     val name: String = "",
     val email: String = "",
     val contact: String = "",
